Updated pedal to 32 bit unsigned value

+++ projects/pedal/inc/pedal_data.h
@@ -3,8 +3,8 @@
 
 // Read the trottle value in mV
 // Requires the MAX11600 to be initialized
-StatusCode read_throttle_data(int16_t *position);
+StatusCode read_throttle_data(uint32_t *position);
 
 // Read the trottle value in mV from the MAX11600
 // Requires the MAX11600 to be initialized
-StatusCode read_brake_data(int16_t *position);
+StatusCode read_brake_data(uint32_t *position);
